Refactored Tests for AISData and FileParser, only cosmetic changes
Created a variable for locations of testfiles, so that it only has to be changed once if the file is moved.

Added nested classes to tests to better organise the tests

Removed a lot of Arrange-Act-Assert Comments. I have left them in the first test, which should make it clear that this is the pattern we use. Therefore i do not think it is needed in every test