Removed threads entirely
All the test pass, and that pretty much all i am going by here. I also ran the program and it works.

But it seemed to me that the thread running CASystem was hardwired to sleep, so as to allow time for the InputSimulator thread to run.
Instead i just made an explicit call to run in the loop of CASystem