Removed unecessary locks, and refactored tests accordingly
There seems to be a little bit of weird behavior sometimes. I have mostly noticed it right after timeframe is set above 0.
Not sure what (if anything) has actually been broken. But it does not seem to cause large problems, or cause problems consistently.

If there is a problem it almost certainly has to do with concurrency.